Group exception and trigger handling tests
