+++ cypress/e2e/case-details/handle-triggers-and-exceptions/edit-fields-with-exceptions/arrestSummonsNumber.cy.ts
