+++ cypress/e2e/case-details/handle-triggers-and-exceptions/edit-fields-with-exceptions/next-hearing-date.cy.ts
@@ -1,7 +1,7 @@
-import nextHearingDateExceptions from "../../../../test/test-data/NextHearingDateExceptions.json"
-import dummyAho from "../../../../test/test-data/error_list_aho.json"
-import hashedPassword from "../../../fixtures/hashedPassword"
-import { verifyUpdatedMessage, submitAndConfirmExceptions } from "../../../support/helpers"
+import nextHearingDateExceptions from "../../../../../test/test-data/NextHearingDateExceptions.json"
+import dummyAho from "../../../../../test/test-data/error_list_aho.json"
+import hashedPassword from "../../../../fixtures/hashedPassword"
+import { verifyUpdatedMessage, submitAndConfirmExceptions } from "../../../../support/helpers"
 
 describe("NextHearingDate", () => {
   before(() => {
