+++ cypress/e2e/case-details/handle-triggers-and-exceptions/edit-fields-with-exceptions/next-hearing-location.cy.ts
@@ -1,7 +1,7 @@
-import dummyAho from "../../../../test/test-data/HO100102_1.json"
-import nextHearingLocationExceptions from "../../../../test/test-data/NextHearingLocationExceptions.json"
-import hashedPassword from "../../../fixtures/hashedPassword"
-import { submitAndConfirmExceptions, verifyUpdatedMessage } from "../../../support/helpers"
+import dummyAho from "../../../../../test/test-data/HO100102_1.json"
+import nextHearingLocationExceptions from "../../../../../test/test-data/NextHearingLocationExceptions.json"
+import hashedPassword from "../../../../fixtures/hashedPassword"
+import { submitAndConfirmExceptions, verifyUpdatedMessage } from "../../../../support/helpers"
 
 describe("NextHearingLocation", () => {
   before(() => {
